+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/impl/Accessor.java
@@ -765,6 +765,8 @@ public abstract class Accessor {
 
         public abstract boolean isSideEffectingTLAction(ThreadLocalAction action);
 
+        public abstract boolean isRecurringTLAction(ThreadLocalAction action);
+
         public abstract void performTLAction(ThreadLocalAction action, ThreadLocalAction.Access access);
 
     }