+++ tools/CHANGELOG.md
@@ -4,8 +4,10 @@ This changelog summarizes major changes between Truffle Tools versions.
 
 ## Version 21.3.0
 * Reimplemented CPUSampler to use the Truffle language safepoints thus deprecating several API functions.
+	* Added new option `--engine.SampleContextInitialization` which includes code executed during context initialization in the general sampling profile instead of grouping it into a single entry.
 * Support for hash interoperability in Insight - no need to use `Truffle::Interop.hash_keys_as_members` anymore
 * [Cooperative heap dumping](https://www.graalvm.org/tools/javadoc/org/graalvm/tools/insight/heap/package-summary.html) when embedding Insight into Java applications
+
 
 ## Version 21.1.0